requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0